+++ src/main/java/day9/Task2/Rectangle.java
@@ -4,7 +4,7 @@ public class Rectangle extends Figure {
     private double height;
     private double width;
 
-    public Rectangle(int height, int weigth, String color) {
+    public Rectangle(double height, double weigth, String color) {
         super(color);
         this.height = height;
         this.width = weigth;
